+++ src/main/java/com/example/zvon73/service/TempleService.java
@@ -4,6 +4,7 @@ import com.example.zvon73.DTO.TempleDto;
 import com.example.zvon73.controller.domain.MessageResponse;
 import com.example.zvon73.controller.domain.TempleOperatorRequest;
 import com.example.zvon73.entity.Bell;
+import com.example.zvon73.entity.Enums.Role;
 import com.example.zvon73.entity.Temple;
 import com.example.zvon73.entity.User;
 import com.example.zvon73.repository.TempleRepository;
@@ -26,22 +27,21 @@ public class TempleService {
     private final TempleRepository templeRepository;
     private final UserService userService;
 
-    private void validate(TempleDto temple){
-        if(temple.getTitle().isEmpty() || temple.getTitle().length() < 5 || temple.getTitle().length() >100)
-            throw new ValidateException("Некорректное название храма");
-        if(temple.getAddress().isEmpty() || temple.getAddress().length() < 5 || temple.getAddress().length() > 100)
-            throw new ValidateException("Некорректный адрес храма");
-        if(temple.getPhone().isEmpty() || temple.getPhone().length() > 12 || temple.getPhone().length() < 11 )
-            throw new ValidateException("Некорректный телефон храма");
-        if(temple.getDescription().isEmpty())
-            throw new ValidateException("Некорректное описание храма");
-        if(temple.getImage().length == 0)
-            throw new ValidateException("Некорректное изображение храма");
+    private boolean checkAdmin(){
+        User currentUser = userService.getCurrentUser();
+        return currentUser.getRole().equals(Role.ADMIN);
+    }
+    private boolean checkRinger(Temple temple){
+        User currentUser = userService.getCurrentUser();
+        return temple.checkRinger(currentUser.getId());
     }
 
+
     @Transactional
-    public Temple create(TempleDto temple)
+    public TempleDto create(TempleDto temple)
     {
+        if(!checkAdmin())
+            return new TempleDto();
         Temple newTemple = Temple.builder()
                 .title(temple.getTitle())
                 .address(temple.getAddress())
@@ -49,43 +49,48 @@ public class TempleService {
                 .description(temple.getDescription())
                 .image(temple.getImage())
                 .build();
-        return templeRepository.save(newTemple);
+        return new TempleDto(templeRepository.save(newTemple));
     }
 
     @Transactional(readOnly = true)
     public Temple findById(UUID id){
         return templeRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException("Храм с данным id не найден"));
+                .orElse(null);
     }
 
     @Transactional
-    public Temple update(TempleDto newTemple)
+    public TempleDto update(TempleDto newTemple)
     {
 
         Temple currentTemple = findById(UUID.fromString(newTemple.getId()));
 
-            if( !currentTemple.getAddress().equals(newTemple.getAddress()) )
-                currentTemple.setAddress(newTemple.getAddress());
+        if(currentTemple ==  null)
+            return new TempleDto();
+        if(!checkAdmin() || !checkRinger(currentTemple))
+            return new TempleDto();
 
-            if( !currentTemple.getPhone().equals(newTemple.getPhone()) )
-                currentTemple.setPhone(newTemple.getPhone());
+        if( !currentTemple.getAddress().equals(newTemple.getAddress()) )
+            currentTemple.setAddress(newTemple.getAddress());
+        if( !currentTemple.getPhone().equals(newTemple.getPhone()) )
+            currentTemple.setPhone(newTemple.getPhone());
+        if( !currentTemple.getTitle().equals(newTemple.getTitle()) )
+            currentTemple.setTitle(newTemple.getTitle());
+        if( !currentTemple.getDescription().equals(newTemple.getDescription()) )
+            currentTemple.setDescription(newTemple.getDescription());
 
-            if( !currentTemple.getTitle().equals(newTemple.getTitle()) )
-                currentTemple.setTitle(newTemple.getTitle());
 
-            if( !currentTemple.getDescription().equals(newTemple.getDescription()) )
-                currentTemple.setDescription(newTemple.getDescription());
-
-            if( Arrays.equals(currentTemple.getImage(), newTemple.getImage()) )
-                currentTemple.setImage(newTemple.getImage());
-
-        return templeRepository.save(currentTemple);
+        return new TempleDto(templeRepository.save(currentTemple));
     }
 
     @Transactional
     public MessageResponse delete(UUID id){
         try {
             Temple currentTemple = findById(id);
+            if(currentTemple ==  null)
+                return new MessageResponse("", "Храм не найден");
+            if(!checkAdmin() || !checkRinger(currentTemple))
+                return new MessageResponse("", "Not Access");
+
             templeRepository.delete(currentTemple);
             return new MessageResponse("Храм успешно удален", "");
         }catch (Exception e){
